requirements.txt
flask
requests
google-generativeai
pymongo
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
python-dotenv
